+++ CHANGELOG.md
@@ -34,6 +34,7 @@ bool AutoCreateNewComplexTypes = true; // Return types of mutations will be adde
 
 - Fix #206 - If a service field included a binary expression EntityGraphQL would sometimes build an incorrect selection expression
 - Fix #214 - The implicit conversion return type of a `ArgumentHelper.Required<>()` field argument is non-null
+- Fix #212 - Regression using a static or instance method in a service field
 
 # 3.0.5